refactor: extract SettingsState and SettingsEvent to separate files
